% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssaySpatialExperiment-spatialMap.R
\name{checkSpatialMap}
\alias{checkSpatialMap}
\title{Check spatialMap consistency}
\usage{
checkSpatialMap(
  spatialMap,
  sampleMap,
  points = PointsLayerList(),
  shapes = ShapesLayerList()
)
}
\arguments{
\item{spatialMap}{A \code{spatialMap} \linkS4class{DataFrame} or \code{NULL}.}

\item{sampleMap}{A MAE \code{sampleMap} \linkS4class{DataFrame}.}

\item{points}{A \linkS4class{PointsLayerList} or named list of point layers.}

\item{shapes}{A \linkS4class{ShapesLayerList} or named list of shape layers.}
}
\value{
Character vector of error messages (length zero if no problems).
}
\description{
Run \code{spatialMap} diagnostics without constructing a
\linkS4class{MultiAssaySpatialExperiment}. Returns a character vector of
problems (empty if valid).
}
