% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssaySpatialExperiment-read-generics.R
\name{readCSVForMASE}
\alias{readCSVForMASE}
\alias{readCSVForMASE,character-method}
\title{Read CSV Files for MASE}
\usage{
readCSVForMASE(file_path, ...)

\S4method{readCSVForMASE}{character}(file_path, ...)
}
\arguments{
\item{file_path}{Character path to CSV file}

\item{...}{Additional arguments passed to format-specific methods}
}
\value{
DataFrame
}
\description{
Read a CSV file into a \linkS4class{DataFrame} for MASE component assembly.
}
