% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssaySpatialExperiment-read-generics.R
\name{readGeoJSONForMASE}
\alias{readGeoJSONForMASE}
\alias{readGeoJSONForMASE,character-method}
\title{Read GeoJSON Files for MASE}
\usage{
readGeoJSONForMASE(file_path, ...)

\S4method{readGeoJSONForMASE}{character}(file_path, quiet = TRUE, ...)
}
\arguments{
\item{file_path}{Character path to GeoJSON file}

\item{...}{Additional arguments passed to format-specific methods}

\item{quiet}{Logical, suppress sf reading messages}
}
\value{
sf or equivalent object
}
\description{
Read a GeoJSON file via \pkg{sf} for shape layers in MASE readers.
}
